Modified typo: return type of getAccountNumber is "int".

+++ src/edu/postech/csed332/homework1/Account.java
@@ -10,7 +10,7 @@ interface Account {
 	 *  
 	 *  @return the account number
 	 */
-	double getAccountNumber();
+	int getAccountNumber();
 	/**
 	 *  Returns the account balance 
 	 *
